% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_hdx_admin.R
\name{download_hdx_adm}
\alias{download_hdx_adm}
\title{Function to download HDX admin boundary}
\usage{
download_hdx_adm(
  country_code,
  admin_level = 3,
  df_index = NULL,
  keep_zip = F,
  output_path = "data-raw"
)
}
\arguments{
\item{country_code}{ISO country code.}

\item{admin_level}{admin level (0-3). Must be numeric. Default is 3.}

\item{df_index}{HDX index number from the printed table. The function should take the index number automatically but some cases (i.e Multiple shapefile), manual entry needed.}

\item{keep_zip}{To keep downloaded shape files (All level of admin boundary will be available under data-raw/OCHA_shapefile folder). Default is FALSE, means, it will delete the folder.}

\item{output_path}{Path where you want to keep the shape files for admin levels}
}
\value{
OCHA common operational data set for specified level as an sf object.
}
\description{
Function to download HDX admin boundary
}
