% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dap_for_dashboard.R
\name{write_dap_for_dashboard}
\alias{write_dap_for_dashboard}
\title{Function to prepare the DAP for dashboard}
\usage{
write_dap_for_dashboard(
  list_of_dataset,
  column_name_for_populaion_group,
  kobo_survey,
  sector_name = c("WASH", "Food Security", "Protection", "Livelihood", "Education",
    "Health", "Nutrition", "Shelter", "Demographic"),
  admin_level = c("admin1", "admin2", "admin3"),
  output_path = "data-raw/data_diagnosis.xlsx"
)
}
\arguments{
\item{list_of_dataset}{A list containing all level of data}

\item{column_name_for_populaion_group}{population group column name}

\item{kobo_survey}{Kobo survey tab}

\item{sector_name}{List of sectors}

\item{admin_level}{List of strata wise admin level}

\item{output_path}{Excel path for the DAP}
}
\value{
Data Analysis Plan for the Dashboard.
}
\description{
Function to prepare the DAP for dashboard
}
