% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/identifying_mismatch_strata_name_with_ocha_cod.R
\name{write_mismatched_strata}
\alias{write_mismatched_strata}
\title{Function write mismatched strata}
\usage{
write_mismatched_strata(
  ocha_cod_level_3 = cntry_adm_3,
  cod_admin1_name = "ADM1_EN",
  cod_admin2_name = "ADM2_EN",
  cod_admin3_name = "ADM3_EN",
  strata_checking_sheet,
  hh_data,
  population_group_column_name_strata_checking = "pop_group",
  output_path = "data-raw/admin_mismatch_fix.xlsx"
)
}
\arguments{
\item{ocha_cod_level_3}{Admin 3 level shape file. Can be downloaded by `download_hdx_adm()`}

\item{cod_admin1_name}{Admin 1 column name}

\item{cod_admin2_name}{Admin 2 column name}

\item{cod_admin3_name}{Admin 3 column name}

\item{strata_checking_sheet}{strata checking sheet, can be found once you run `write_dap_for_dashboard()`}

\item{hh_data}{HH data}

\item{population_group_column_name_strata_checking}{Column name in `hh_data` containing population group information.}

\item{output_path}{Output path}
}
\value{
Write mismatched strata
}
\description{
Function write mismatched strata
}
