% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_cod.R
\name{write_renamed_ocha_admin_name}
\alias{write_renamed_ocha_admin_name}
\title{Function to write renamed OCHA boundaries which would align with dataset}
\usage{
write_renamed_ocha_admin_name(
  ocha_cod_level_3 = cntry_adm_3,
  cod_admin1_name,
  cod_admin2_name,
  cod_admin3_name,
  strata_checking_sheet,
  mistached_list,
  output_path = "data-raw/OCHA_renamed_adm3.shp"
)
}
\arguments{
\item{ocha_cod_level_3}{Admin 3 level shape file. Can be downloaded by `download_hdx_adm()`}

\item{cod_admin1_name}{Admin 1 column name}

\item{cod_admin2_name}{Admin 2 column name}

\item{cod_admin3_name}{Admin 3 column name}

\item{strata_checking_sheet}{strata checking sheet, can be found once you run `write_dap_for_dashboard()`}

\item{mistached_list}{Filled output from `write_mismatched_strata()`}

\item{output_path}{Output path}
}
\value{
Write mismatched strata
}
\description{
Function to write renamed OCHA boundaries which would align with dataset
}
